Example test of divergent delta table snapshots in one query
This is a test which illustrates the consequences of the multiple snapshots loads found in https://github.com/trinodb/trino/issues/16782

The test uses two delta lake catalogs, where one reads and the other writes. We add a hook to the TrinoFileSystem newInputFile calls which allows us to make new commits at appropriate times.

We run the query `SELECT COUNT(*) FROM delta_table UNION SELECT COUNT(*) FROM delta_table`. To determine which parquet files to read, a TableSnapshot is loaded once for each of the two DeltaLakeTableHandles handles in the above query in the DeltaLakeSplitManager. We use the hook to add a commit right before the TableSnapshot is loaded in the DeltaLakeSplitManager for the second DeltaLakeTableHandle. This means that the table handles use different snapshots, and arrive at the row counts 1 or 2, respectively. Hence, the result set of the union of the count of the same table twice has two rows, 1 and 2.

In a similar way it should also be possible to trigger a crash when reading data if the table schema changes in a breaking way between the analysis phase, which determines the DeltaLakeColumnHandles, and the reading of the data using the column handles in the DeltaLakeSplitManager. As as far as I know, we cannot make breaking changes to the schema of a Delta table with Trino, so I was not able to write such a test case. Other writes could make breaking changes to the table schema (e.g., drop a column), so such crashes should be possible.